refactor: ipfs provider config
